Add files via upload
This Projects Online Library Management System introduces a System that is created by using the latest tech stacks i.e. HTML5, Bootstrap, and JavaScript(ES6)

One who has a basic understanding of JavaScript and familiar with JS Functions concepts can understand easily. 

You can customize this project’s animation as well as logic as per requirement.

It’s one of the major projects you can use for your personal use as well as to present in your college.

This is about a Library which manages the records of the books its author and the type of book you want to store.